% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{gobayes_get_server}
\alias{gobayes_get_server}
\title{Get Current GoBayes Server URL}
\usage{
gobayes_get_server(masked = TRUE)
}
\arguments{
\item{masked}{Logical. Whether to mask the URL in output. Default TRUE.}
}
\value{
Character. Current server URL.
}
\description{
Get Current GoBayes Server URL
}
