% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{gobayes_health_check}
\alias{gobayes_health_check}
\title{Health Check for GoBayes Server}
\usage{
gobayes_health_check()
}
\value{
Invisibly returns TRUE if the server is reachable and healthy.
}
\description{
Health Check for GoBayes Server
}
