% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{gobayes_set_api_token}
\alias{gobayes_set_api_token}
\title{Set GoBayes API Token}
\usage{
gobayes_set_api_token(token)
}
\arguments{
\item{token}{Character. API token issued by the GoBayes service.}
}
\value{
Invisibly returns TRUE.
}
\description{
Set GoBayes API Token
}
